+++ Flow/Changes.md
@@ -37,6 +37,7 @@
 - Added a **Sided[A]** type alias for **Either[A, A]** to **CollectionExtensions**
 - Added a **Mutate[A]** type alias for **A => A** functions
 - Added a couple new constructor for span classes
+- **Pair**`.separateMatching` now uses `EqualsFunction.default` by default
 
 ## v2.1 - 01.05.2023
 This update mostly introduces new utility functions and some new utility classes.